chore(suite): move stakingPools to account.misc
