+++ suite-common/wallet-types/src/account.ts
@@ -1,7 +1,9 @@
 import { Network, BackendType, NetworkSymbol } from '@suite-common/wallet-config';
 import { AccountEntityKeys } from '@suite-common/metadata-types';
 import { AccountInfo, PROTO, TokenInfo } from '@trezor/connect';
-import { StakingPool } from '@trezor/blockchain-link-types';
+import {
+    StakingPool,
+} from '@trezor/blockchain-link-types/src/blockbook-api';
 
 export type MetadataItem = string;
 export type XpubAddress = string;
@@ -42,7 +44,10 @@ type AccountNetworkSpecific =
       }
     | {
           networkType: 'ethereum';
-          misc: { nonce: string };
+          misc: {
+              nonce: string;
+              stakingPools?: StakingPool[];
+          };
           marker: undefined;
           page: AccountInfo['page'];
       }
@@ -93,7 +98,6 @@ export type Account = {
      * metadata/labeling feature which requires device for encryption. local accountLabel field was introduced.
      */
     accountLabel?: string;
-    stakingPools?: StakingPool[];
 } & AccountBackendSpecific &
     AccountNetworkSpecific;
